dataflow/streaming_opencv/script/requirements.txt
google-cloud-storage
numpy
